Centralising Map Component
Moving the major marker and GeoJSON layers on the map to the central component to ease future modifications.
